feat: 게임 종료 메뉴 구현
